The following regression tests failures occur on PPC64 little endian only.
The regression test none/tests/jm_vec/isa_2_07 has failures on the lxsiwax and
lxsiwzx instructions. They are loads and the the results are correct for
big endian but not little endian. The little endian result matches the
expected big endian result. 

The regresssion test none/tests/test_isa_2_07_part2 has a failure with the
vbpermq instruction. The little endian result matches the expected result for
big endian. The upper and lower 64 bits of the result are not swapped correctly
for little endian.

This commit fixes these issues.

The bugzilla for the issue is 346270.

git-svn-id: svn://svn.valgrind.org/vex/trunk@3134 8f6e269a-dfd6-0310-a8e1-e2731360e62c